refactor: Phase 5 utility refactoring - move tag validation to utils
Move validate_and_normalize_tags out of the repository layer into a
dedicated utility module, following the principle that pure business
validation logic should not live in data-access code.

New file: src/utils/tags.rs
- validate_and_normalize_tags()

Updated src/utils/mod.rs:
- Added `pub mod tags` declaration
- Added pub use tags::validate_and_normalize_tagsi re-export

Updated call sites (import changed from tag_repository to crate::utils):
- src/api/links/create.rs
- src/api/links/update.rs
- src/api/links/import.rs
- src/services/link_service.rs (inline path updated)

Updated src/repositories/tag_repository.rs:
- Removed validate_and_normalize_tags function
- Removed now-unused crate::utils::normalize_tag import
