+++ src/repositories/tag_repository.rs
@@ -1,7 +1,6 @@
 /// Tag repository - Data access for tags
 ///
 /// Handles all database operations related to tags.
-use crate::utils::normalize_tag;
 use worker::d1::D1Database;
 use worker::*;
 
@@ -10,30 +9,6 @@ use worker::*;
 pub struct OrgTag {
     pub name: String,
     pub count: i64,
-}
-
-/// Validate and normalize a list of tags. Returns an error if any tag is invalid.
-pub fn validate_and_normalize_tags(tags: &[String]) -> Result<Vec<String>> {
-    if tags.len() > 20 {
-        return Err(worker::Error::RustError(
-            "Maximum 20 tags per link".to_string(),
-        ));
-    }
-    let mut normalized = Vec::with_capacity(tags.len());
-    for tag in tags {
-        match normalize_tag(tag) {
-            Some(t) => normalized.push(t),
-            None => {
-                return Err(worker::Error::RustError(format!(
-                    "Invalid tag: '{}'. Tags must be non-empty and at most 50 characters.",
-                    tag
-                )));
-            }
-        }
-    }
-    let mut seen = std::collections::HashSet::new();
-    normalized.retain(|t| seen.insert(t.to_lowercase()));
-    Ok(normalized)
 }
 
 /// Repository for tag operations
