+++ src/utils/mod.rs
@@ -5,6 +5,7 @@ pub mod errors;
 pub mod http;
 pub mod query_params;
 pub mod short_code;
+pub mod tags;
 pub mod time;
 pub mod url;
 pub mod url_normalization;
@@ -16,6 +17,7 @@ pub use errors::AppError;
 pub use http::{get_client_ip, hash_ip};
 pub use query_params::QueryParams;
 pub use short_code::{generate_short_code, generate_short_code_with_length};
+pub use tags::validate_and_normalize_tags;
 pub use time::now_timestamp;
 pub use url_normalization::normalize_url_for_blacklist;
 pub use validation::{normalize_tag, validate_short_code, validate_url};
